LF-5388 Move functions to utils file
