+++ packages/webapp/src/containers/Profile/FarmSettings/MarketDirectory/Certifications/index.tsx
@@ -16,11 +16,9 @@
 import clsx from 'clsx';
 import { useHistory } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
-import { toCertificationItems } from '../../../../Certifications/utils';
+import { getCertificationStatus, toCertificationItems } from '../../../../Certifications/utils';
 import CertificationsEmptyState from '../../../../../components/Certifications/CertificationsEmptyState';
-import CertificationCard, {
-  getCertificationStatus,
-} from '../../../../../components/Certifications/CertificationCard';
+import CertificationCard from '../../../../../components/Certifications/CertificationCard';
 import {
   Certification,
   SupportedCertificationSystemType,
